requirements.txt
pandas=0.23.0
pdfminer.six=20170720
lxml=4.1.1
#textract # requires swig
